requirements.txt
# Core dependencies
tensorflow>=2.12.0
numpy>=1.24.0
Pillow>=9.5.0
h5py>=3.9.0

# GUI dependencies
customtkinter>=5.2.0
darkdetect>=0.7.1
tk>=0.1.0

# Web application
Flask>=2.3.0

# Visualization and data analysis
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Image processing
opencv-python>=4.8.0

# Development and packaging tools
pyinstaller>=5.13.0
pytest>=7.4.0

# Progress bars
tqdm>=4.66.0

# Optional dependencies that might be needed
# for specific functionality or future development
requests>=2.31.0
